Add support for a builder with required components
When the record components are annotated with `@RecordBuilder.Required`,
the builder method will take the required record components as its
parameters. This is particularly useful to ensure that the builder is
always populated with the required fields at compile-time instead of
relying on a runtime validation.

Example:
```
public record MyRecord(@Required int a, @Required int b, int c)

@Generated("io.soabase.recordbuilder.core.RecordBuilder")
public class MyRecordBuilder {
    ...
    @Generated("io.soabase.recordbuilder.core.RecordBuilder")
    public static MyRecordBuilder builder(int a, int b) {
        return new MyRecordBuilder().a(a).b(b);
    }
    ...
}
```