Add missing annotations to builder all-args constructor
Motivation
----------

I am working with record-builder and integrating tools like NullAway. Everything was working great, but I noticed that the generated builder all-args constructor doesn't include the annotations found on the record. This causes tools like NullAway to fail when inspecting the builder as a nullable accessor could be trying to set a value on to a non-nullable constructor argument.

Modifications
-------------

The fix is relatively simple. There's already an addConstructorAnnotations method designed to perform this action. It was not being used for the all-args constructor. I've added its usage and the annotations started appearing.

Result
------

The result is the generated builder all-args constructor mirrors the annotations found on the record