Accessor annotations must be supplemented by type annotations
It appears that calling `component.getAccessor().getAnnotationMirrors()`
on a record component accessor does not return all annotations for
some reason. This was determined by debugging and trial-and-error.

The annotations returned for the accessor must be supplemented with
annotations from the `asType()` called on the component. Note: the
annotations from `asType()` are not complete either! It must be a
union of both.

Fixes #111