Don't add annotation if it doesn't have the correct `@Target`
This is a further refinement of the `TypeUse` annotation issue. In this
case when we copy annotations from the record, the target of the annotation
will be different than `ElementType.RECORD_COMPONENT` or
`ElementType.TYPE_USE`, etc. Ensure that the annotation has a compatible
`@Target` when adding.

Added a test record for this as well provided by @facboy

Fixes #258